adb: Retry io in case of interrupt
io_submit sleeps waiting for the gadget
to be enabled. Currently if that sleep
is interrupted it will shut down adb,
causing it to have to start back up again.

Rather than return EINTR if interrupted,
io_submit actually completes and the EINTR
is found later when looking through events.
Since an io that is interrupted will be
small anyway, add a loop to retry small ios.

Also upgrade aio logs in accordance with
their importance.

Fixes: 75981904
Test: adb works, logs show successful interrupt handling
Change-Id: I35973fce130ee849ce59fef80d15b65afb816ba4

Former-commit-id: 81b0254e182da4f6583f96c22a616cba721a476f [formerly 97369224a234273d468ecffdf4c7e640e5737bc6] [formerly 97369224a234273d468ecffdf4c7e640e5737bc6 [formerly 29add682044b826015f1388f528308310a0d60fe]] [formerly ae545b437b0084a8347584301e7473a318eb3510 [formerly ae545b437b0084a8347584301e7473a318eb3510 [formerly 6e9a327504e6aa2488b736fd5592525278a0916f]]]
Former-commit-id: ae545b437b0084a8347584301e7473a318eb3510
Former-commit-id: 9527270c898ddca0b8482e2742dfb032e13116ed [formerly e2fb6a4bd500da01b61b29738cb0d217e31a0978]
Former-commit-id: 83460767d58146c3e2fbdf5fe8fbc3ee5a19eaef