Fix null pointer dereference in RegsArm.
Fix RegsArm::GetPcAdjustment to check for an invalid elf before trying
to read memory.
Modify the tests for this so it crashes without this change.

Also modify the GetPcAdjustment for all different architectures so
that unless the relative pc is too small, it will return the minimum
amount that should be adjusted. This is to handle cases where we still
want to adjust the pc but it's in an invalid elf. Mostly this is for
handling cases when the pc is in jit gdb debug code so that we use the
right unwind information.

Bug: 77233204

Test: Passes unit tests for libbacktrace/libunwindstack.
Change-Id: Id73609adaf3b80a583584441de228156fec3afa7
(cherry picked from commit a62f988495ed209909e3e7a27758effdae684f96 [formerly a62f988495ed209909e3e7a27758effdae684f96 [formerly a62f988495ed209909e3e7a27758effdae684f96 [formerly 6dbc28ece3ab7cadd0087b4dc31ba9a2986545f0]]])

Former-commit-id: 61d65bf95f77a85d3e99e9ba6e6d84fd3e20b2e4 [formerly 61d65bf95f77a85d3e99e9ba6e6d84fd3e20b2e4 [formerly 939163b3c06db9ef804bd8aab9477491a69266e3]] [formerly 860b52b910aaa03d32b55bd4f19e6a21d36cf067 [formerly 860b52b910aaa03d32b55bd4f19e6a21d36cf067 [formerly fd6b7284ee4154bdab11da7d332920dc8b6aa342]]]
Former-commit-id: 860b52b910aaa03d32b55bd4f19e6a21d36cf067
Former-commit-id: c1196b04f8f760cab60330ecb0d9d2c2ff57bb1d [formerly 96b55c7d049e0009081f537dbf8f5d99d5d0a787]
Former-commit-id: 58d6a1eccc8167ffa0d7e1d82aa6eadde8168c20