Make logd more aggressive when scanning for the position from which to resume logging.
Events in the LogBuffer are supposed to be sorted by timestamp, but for a variety
of reasons that doesn't always happen.  When a LogReader is reading from LogBuffer,
LogBuffer starts at the newest event, and scans backward through the list, looking
for the last event.  Previously it would accept a couple that were a little bit out
of order, but if it found one that was ancient, it would just bail. This change
removes that check for the ancient messages.  They are probably indicative of
something else upstream, but since there is no invariant of the list being sorted,
this change simplifies the search algorithm, and makes it look only at the previous
300 events.

Bug: 77222120
Test: while true ; do frameworks/base/cmds/statsd/run_tests.sh 2h ; done
Change-Id: I0824ee7590d34056ce27233a87cd7802c28f50e4

Former-commit-id: 2684773cde790e9ea26223369fe2e99e4e1c41ba [formerly 87cc1446d77007562b614ec8803418b0a033a483] [formerly 87cc1446d77007562b614ec8803418b0a033a483 [formerly c6ae77fad9cfa5794100c8fbf32a6720ab3568af]] [formerly 096aae1df8db0496b9d8f4302667540ae5316e03 [formerly 096aae1df8db0496b9d8f4302667540ae5316e03 [formerly 4bba6982455de327bdf018c386032277a08860d5]]]
Former-commit-id: 096aae1df8db0496b9d8f4302667540ae5316e03
Former-commit-id: 3a3c3b1dc029327acf07d17fccdcefb40c7eac92 [formerly b7f5fe292587951b30807372a96057c761b4b33f]
Former-commit-id: 964341cae085e7ef217d39b357fc81e93bf13ad5